FilmFlare/MiniProject/MiniProject1/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "appwrite>=10.0.0",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "google-generativeai>=0.8.5",
    "requests>=2.32.3",
]
